added prefixes to files updated tests created tests and scripts for worms requests
