--- files/raw/user_profile.json
@@ -1,27 +1,0 @@
-[
-    [
-        {
-            "id": 4153234,
-            "login": "d_gonzalez",
-            "spam": false,
-            "suspended": false,
-            "created_at": "2021-03-29T07:52:40+00:00",
-            "login_autocomplete": "d_gonzalez",
-            "login_exact": "d_gonzalez",
-            "name": "Daniel González-Suárez",
-            "name_autocomplete": "Daniel González-Suárez",
-            "orcid": null,
-            "icon": "https://static.inaturalist.org/attachments/users/icons/4153234/thumb.jpeg?1724183967",
-            "observations_count": 349,
-            "identifications_count": 17,
-            "journal_posts_count": 0,
-            "activity_count": 366,
-            "species_count": 159,
-            "annotated_observations_count": 4,
-            "universal_search_rank": 349,
-            "roles": [],
-            "site_id": 17,
-            "icon_url": "https://static.inaturalist.org/attachments/users/icons/4153234/medium.jpeg?1724183967"
-        }
-    ]
-]
